+++ examples/count_hands.rs
@@ -1,0 +1,14 @@
+use leaprs::*;
+
+fn main() {
+    let mut c = Connection::create(ConnectionConfig::default()).unwrap();
+    c.open().unwrap();
+    loop {
+        if let Ok(msg) = c.poll(1000) {
+            match msg.event() {
+                Event::Tracking(e) => println!("{} hand(s)", e.hands().len()),
+                _ => {}
+            }
+        }
+    }
+}