+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ### Changed
 
+- Improved the examples to avoid common timeout unwrap()
+
 ### Removed
 
 - **Breaking** All the trivial accessor functions were removed, in favour of the `Deref` trait.